chore(utxo-core): generated index.ts and moved test function
TICKET: BTC-2150
