+++ modules/utxo-core/src/index.ts
@@ -1,7 +1,7 @@
 export * as bip65 from './bip65';
 export * as descriptor from './descriptor';
 export * as testutil from './testutil';
+export * as paygo from './paygo';
 export * from './dustThreshold';
 export * from './Output';
 export * from './xOnlyPubkey';
-export * from './paygo/ExtractAddressPayGoAttestation';
