+++ modules/utxo-core/src/paygo/index.ts
@@ -1,0 +1,1 @@
+export * from './ExtractAddressPayGoAttestation';
