+++ modules/utxo-core/src/testutil/index.ts
@@ -1,3 +1,4 @@
 export * from './fixtures.utils';
 export * from './key.utils';
 export * from './toPlainObject.utils';
+export * from './generatePayGoAttestationProof.utils';
